requirements.txt
Django==2.0.6
djangorestframework==3.10.0
et-xmlfile==1.0.1
jdcal==1.4.1
numpy==1.20.1
openpyxl==3.0.6
pandas==1.2.2
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
xlrd==2.0.1
uwsgi
